+++ NEWS.md
@@ -2,6 +2,10 @@
 
 # teal.picks 0.2.0
 
+### Bug fixes
+
+* Fixed a bug that crashes the `picks_srv` when datasets with labels that contain lists. Defaults to the choice name (#96).
+
 ### Miscellaneous
 
 * `variables()` allows for custom `allow-clear` options to be set by user without being overwritten.