+++ NEWS.md
@@ -1,4 +1,4 @@
-# teal.picks 0.2.0.9003
+# teal.picks 0.2.0.9004
 
 ### Enhancements